Create README.md
General steps for creating 3D models from segmentation. It was difficult for Dylan to map the coordinates of the synapse objects back to the original models from the paper so the models (both neurons and synapses) were re-made such that each object to retain their relative position when exporting. This avoided the problem of translating CATMAID/VAST synapse positions onto 3d models.